Images For my Razorpay clone website
The aim of this project was to create a clone website of Razorpay, a popular online payment gateway, using Tailwind CSS as the front-end framework. The objective was to replicate the key features and visual design of Razorpay's website while leveraging the capabilities of Tailwind CSS for efficient and responsive web development.